split salmon and added kallisto

+++ scripts/das_suppa2.sh
@@ -16,7 +16,7 @@ if [[ ${PIPESTATUS[0]} -ne 4 ]]; then
 fi
 
 OPTIONS=
-LONGOPTS=pdata:,gtf:,out:,nthread:,log:
+LONGOPTS=pdata:,gtf:,out:,nthread:,log:,salmon,salmonstar,kallisto
 
 # -regarding ! and PIPESTATUS see above
 # -temporarily store output to be able to check for errors
@@ -31,7 +31,7 @@ fi
 # read getopt’s output this way to handle the quoting right:
 eval set -- "$PARSED"
 
-pdata=- out=- gtf=- nthread=4
+pdata=- out=- gtf=- nthread=4 salmon=n salmonstar=n kallisto=n
 # now enjoy the options in order and nicely split until we see --
 while true; do
     case "$1" in
@@ -50,6 +50,18 @@ while true; do
         --nthread)
         	nthread="$2"
             shift 2
+            ;;
+        --kallisto)
+            kallisto=y
+            shift
+            ;;
+        --salmon)
+            salmon=y
+            shift
+            ;;
+        --salmonstar)
+            salmonstar=y
+            shift
             ;;
 		--log)
         	log="$2"
@@ -80,31 +92,27 @@ mkdir -p $out/SUPPA2
 #extract events from gtf
 $SUPPA generateEvents -i $gtf -o $out/SUPPA2/$GTFNAME -f ioi
 
-for d in "READS" "STAR"; do
+if [[ "$salmon" = "y" ]]; then
+	d="READS"
 	root=$out/SALMON/$d
 	for cond in `sed '1d' $pdata | cut -f2 | sort -u`; do
 		files=`grep -P "\t$cond" $pdata | cut -f1`
 		f=`for file in $files; do echo $root/$file/quant.sf; done`
 		
-		watch pidstat -dru -hlH '>>' $log/suppa2_psiPeform_$cond-$(date +%s).pidstat & wid=$!
+		watch pidstat -dru -hlH '>>' $log/suppa2_psiPeform_salmon_${d}_$cond-$(date +%s).pidstat & wid=$!
 
 		## salmon is -k1 -f4; for kallisto -k1 -f5
 		python3 /home/software/SUPPA//multipleFieldSelection.py \
-			-i $f -k 1 -f 4 -o $out/SUPPA2/salmon_$d.$cond.tmp.counts
+			-i $f -k 1 -f 4 -o $out/SUPPA2/salmon_$cond.$d.tmp.counts
 
 		python3 /home/software/SUPPA//multipleFieldSelection.py \
-			-i $f -k 1 -f 5 -o $out/SUPPA2/salmon_$d.$cond.read.counts
+			-i $f -k 1 -f 5 -o $out/SUPPA2/salmon_$cond.$d.read.counts
 		
 		$SUPPA psiPerIsoform -g $gtf -e $out/SUPPA2/salmon_$(basename $root).$cond.tmp.counts -o $out/SUPPA2/$d.$cond &> /dev/null
 
 		kill -15 $wid
 	done
 
-#python3.4 suppa.py joinFiles -f tpm -i sample1.tpm sample2.tpm sample3.tpm -o all_samples_tpms
-
-##need to mute these 2 calls
-#$SUPPA psiPerIsoform -g $gtf -e $out/COUNTS/tpm.counts.0 -o $out/SUPPA2/c1 &> /dev/null
-#$SUPPA psiPerIsoform -g $gtf -e $out/COUNTS/tpm.counts.1 -o $out/SUPPA2/c2 &> /dev/null
 	watch pidstat -dru -hlH '>>' $log/suppa2_diff_$d-$(date +%s).pidstat & wid=$!
 
 	( [ -f "$out/diff_splicing_outs/SUPPA_salmon_$d.out" ] && echo "[INFO] [SUPPA2] $out/diff_splicing_outs/SUPPA_salmon_$d.out already exists, skipping.."$'\n' ) \
@@ -113,9 +121,74 @@ for d in "READS" "STAR"; do
 
 	kill -15 $wid
 
-#cat $outDir/SUPPA2/SUPPA.out.dpsi.temp.0 | awk '{split($1,a,";"); print a[1]"\t"$2"\t"$3}' > $outDir/SUPPA2/test_suppa.out
 	( [ -f "$out/diff_splicing_outs/SUPPA_salmon_$d.out" ]) || \
 	(cat $out/SUPPA2/SUPPA_salmon_$d.out.dpsi | awk '{split($1,a,";"); print a[1]"\t"$2"\t"$3}' > $out/diff_splicing_outs/SUPPA_salmon_$d.out)
-done
-#mv $OUTDIR/SUPPA.out.dpsi $5
-#rm -r $OUTDIR
+fi
+
+
+if [[ "$salmonstar" = "y" ]]; then
+	d="STAR"
+	root=$out/SALMON/$d
+	for cond in `sed '1d' $pdata | cut -f2 | sort -u`; do
+		files=`grep -P "\t$cond" $pdata | cut -f1`
+		f=`for file in $files; do echo $root/$file/quant.sf; done`
+		
+		watch pidstat -dru -hlH '>>' $log/suppa2_psiPeform_salmon_${d}_$cond-$(date +%s).pidstat & wid=$!
+
+		## salmon is -k1 -f4; for kallisto -k1 -f5
+		python3 /home/software/SUPPA//multipleFieldSelection.py \
+			-i $f -k 1 -f 4 -o $out/SUPPA2/salmon_$cond.$d.tmp.counts
+
+		python3 /home/software/SUPPA//multipleFieldSelection.py \
+			-i $f -k 1 -f 5 -o $out/SUPPA2/salmon_$cond.$d.read.counts
+		
+		$SUPPA psiPerIsoform -g $gtf -e $out/SUPPA2/salmon_$(basename $root).$cond.tmp.counts -o $out/SUPPA2/$d.$cond &> /dev/null
+
+		kill -15 $wid
+	done
+
+	watch pidstat -dru -hlH '>>' $log/suppa2_diff_$d-$(date +%s).pidstat & wid=$!
+
+	( [ -f "$out/diff_splicing_outs/SUPPA_salmon_$d.out" ] && echo "[INFO] [SUPPA2] $out/diff_splicing_outs/SUPPA_salmon_$d.out already exists, skipping.."$'\n' ) \
+			|| ($SUPPA diffSplice -m empirical --input $out/SUPPA2/$GTFNAME.ioi --psi $out/SUPPA2/$d*_isoform.psi \
+    	-e $out/SUPPA2/salmon_$d.*.tmp.counts -gc -o $out/SUPPA2/SUPPA_salmon_$d.out)
+
+	kill -15 $wid
+
+	( [ -f "$out/diff_splicing_outs/SUPPA_salmon_$d.out" ]) || \
+	(cat $out/SUPPA2/SUPPA_salmon_$d.out.dpsi | awk '{split($1,a,";"); print a[1]"\t"$2"\t"$3}' > $out/diff_splicing_outs/SUPPA_salmon_$d.out)
+fi
+
+
+if [[ "$kallisto" = "y" ]]; then
+	d="quant"
+	root=$out/KALLISTO/$d
+	for cond in `sed '1d' $pdata | cut -f2 | sort -u`; do
+		files=`grep -P "\t$cond" $pdata | cut -f1`
+		f=`for file in $files; do echo $root/$file/abundance.tsv; done`
+		
+		watch pidstat -dru -hlH '>>' $log/suppa2_psiPeform_kallisto_$cond-$(date +%s).pidstat & wid=$!
+
+		## salmon is -k1 -f4; for kallisto -k1 -f5
+		python3 /home/software/SUPPA//multipleFieldSelection.py \
+			-i $f -k 1 -f 5 -o $out/SUPPA2/kallisto_$cond.tmp.counts
+
+		python3 /home/software/SUPPA//multipleFieldSelection.py \
+			-i $f -k 1 -f 4 -o $out/SUPPA2/kallisto_$cond.read.counts
+		
+		$SUPPA psiPerIsoform -g $gtf -e $out/SUPPA2/kallisto_$cond.tmp.counts -o $out/SUPPA2/kallisto.$cond &> /dev/null
+
+		kill -15 $wid
+	done
+
+	watch pidstat -dru -hlH '>>' $log/suppa2_diff_$d-$(date +%s).pidstat & wid=$!
+
+	( [ -f "$out/diff_splicing_outs/SUPPA_kallisto.out" ] && echo "[INFO] [SUPPA2] $out/diff_splicing_outs/SUPPA_kallisto.out already exists, skipping.."$'\n' ) \
+			|| ($SUPPA diffSplice -m empirical --input $out/SUPPA2/$GTFNAME.ioi --psi $out/SUPPA2/kallisto*_isoform.psi \
+    	-e $out/SUPPA2/kallisto*.tmp.counts -gc -o $out/SUPPA2/SUPPA_kallisto.out)
+
+	kill -15 $wid
+
+	( [ -f "$out/diff_splicing_outs/SUPPA_kallisto.out" ]) || \
+	(cat $out/SUPPA2/SUPPA_kallisto.out.dpsi | awk '{split($1,a,";"); print a[1]"\t"$2"\t"$3}' > $out/diff_splicing_outs/SUPPA_kallisto.out)
+fi
